+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Version 1.6.0:
 
 * Added a better message when a `CanReleaseResource` implicit evidence
   parameter cannot be found for `grizzled.util.withCloseable()`.
+* Added `GrizzledString.escapeNonPrintables` and `GrizzledChar.isPrintable`
+  methods.
 * Added `toTry` method to the rich `Either` class, to convert an
   `Either` to a `Try`.